change proto types

+++ types/test_util.go
@@ -4,7 +4,7 @@ import (
 	"time"
 )
 
-func MakeCommit(blockID BlockID, height int64, round int,
+func MakeCommit(blockID BlockID, height int64, round int32,
 	voteSet *VoteSet, validators []PrivValidator, now time.Time) (*Commit, error) {
 
 	// all sign
@@ -12,7 +12,7 @@ func MakeCommit(blockID BlockID, height int64, round int,
 		addr := validators[i].GetPubKey().Address()
 		vote := &Vote{
 			ValidatorAddress: addr,
-			ValidatorIndex:   i,
+			ValidatorIndex:   int32(i),
 			Height:           height,
 			Round:            round,
 			Type:             PrecommitType,
